+++ packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/file/directoryPath.ts
@@ -1,6 +1,6 @@
 import { posix } from 'path';
 import { assertNotUndefined } from '../../../package-agnostic-utilities/nil/assertNotUndefined';
-import { buildNamedConstructorFunction } from '../../../package-agnostic-utilities/constructor-function/namedConstructorFunctionBuilder';
+import { buildNamedConstructorFunction } from '../../../package-agnostic-utilities/constructor-function/buildNamedConstructorFunction';
 import { SimplifyN } from '../../../package-agnostic-utilities/type/simplify';
 import { getFileSystemNodePathPartList } from './getFileSystemNodePathPartList';
 import { NodePathConstructorInput } from './nodePathConstructorInput';
