+++ packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/file/file.ts
@@ -1,4 +1,4 @@
-import { buildNamedConstructorFunction } from '../../../package-agnostic-utilities/constructor-function/namedConstructorFunctionBuilder';
+import { buildNamedConstructorFunction } from '../../../package-agnostic-utilities/constructor-function/buildNamedConstructorFunction';
 import { FileExtensionSuffixIdentifier } from '../../../package-agnostic-utilities/file/fileExtensionSuffixIdentifier';
 import { FilePath, FilePathInstance } from './filePath';
 import { FileSystemNodeZorn } from './fileSystemNode';
