+++ packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/linting/lintAssertionOmission.ts
@@ -1,7 +1,7 @@
 import { AbstractInMemoryVoictent } from '../../../layer-agnostic-utilities/voictent/abstractInMemoryVoictent';
 import { InMemoryIndexByName } from '../../../layer-agnostic-utilities/voque/inMemoryVoque';
 import { Voque } from '../../../core/types/voque/voque';
-import { buildNamedConstructorFunction } from '../../../package-agnostic-utilities/constructor-function/namedConstructorFunctionBuilder';
+import { buildNamedConstructorFunction } from '../../../package-agnostic-utilities/constructor-function/buildNamedConstructorFunction';
 import {
   GenericZorn2Template,
   Zorn2,
