+++ packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/linting/rule.ts
@@ -1,4 +1,4 @@
-import { buildNamedConstructorFunction } from '../../../package-agnostic-utilities/constructor-function/namedConstructorFunctionBuilder';
+import { buildNamedConstructorFunction } from '../../../package-agnostic-utilities/constructor-function/buildNamedConstructorFunction';
 import {
   GenericZorn2Template,
   Zorn2,
