+++ packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/type-script-file/comment/adapted/adaptedNormalBlockComment.ts
@@ -1,5 +1,5 @@
 import { TSESTree } from '@typescript-eslint/typescript-estree';
-import { buildNamedConstructorFunction } from '../../../../../package-agnostic-utilities/constructor-function/namedConstructorFunctionBuilder';
+import { buildNamedConstructorFunction } from '../../../../../package-agnostic-utilities/constructor-function/buildNamedConstructorFunction';
 import { AdaptedCommentTypeName } from './adaptedCommentTypeName';
 import { CommentZorn } from '../commentZorn';
