+++ packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/type-script-file/comment/categorized/descriptiveBlockComment.ts
@@ -1,4 +1,4 @@
-import { buildNamedConstructorFunction } from '../../../../../package-agnostic-utilities/constructor-function/namedConstructorFunctionBuilder';
+import { buildNamedConstructorFunction } from '../../../../../package-agnostic-utilities/constructor-function/buildNamedConstructorFunction';
 import { AdaptedJsdocLikeBlockComment } from '../adapted/adaptedJsdocLikeBlockComment';
 import { CommentZorn } from '../commentZorn';
 import { CategorizedCommentTypeName } from './categorizedCommentTypeName';
