+++ packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/type-script-file/comment/categorized/eslintDirectiveComment.ts
@@ -1,4 +1,4 @@
-import { buildNamedConstructorFunction } from '../../../../../package-agnostic-utilities/constructor-function/namedConstructorFunctionBuilder';
+import { buildNamedConstructorFunction } from '../../../../../package-agnostic-utilities/constructor-function/buildNamedConstructorFunction';
 import { AdaptedLineComment } from '../adapted/adaptedLineComment';
 import { AdaptedNormalBlockComment } from '../adapted/adaptedNormalBlockComment';
 import { CategorizedCommentTypeName } from './categorizedCommentTypeName';
