+++ packages/voictents-and-estinants-engine/src/adapted-programs/programs/render-knowledge-graph/boundary/boundaryFact.ts
@@ -1,5 +1,5 @@
 import { InMemoryOdeshin2ListVoque } from '../../../../layer-agnostic-utilities/voictent/inMemoryOdeshinVoictent2';
-import { buildNamedConstructorFunction } from '../../../../package-agnostic-utilities/constructor-function/namedConstructorFunctionBuilder';
+import { buildNamedConstructorFunction } from '../../../../package-agnostic-utilities/constructor-function/buildNamedConstructorFunction';
 import { SimplifyN } from '../../../../package-agnostic-utilities/type/simplify';
 import { Metadata } from '../app/browser/dynamicComponentTypes';
 import { CommonBoundaryRoot } from '../common-boundary-root/commonBoundaryRoot';
