+++ packages/voictents-and-estinants-engine/src/adapted-programs/programs/render-knowledge-graph/dependency/dependency-path/fileDependencyPathSegment.ts
@@ -1,4 +1,4 @@
-import { buildNamedConstructorFunction } from '../../../../../package-agnostic-utilities/constructor-function/namedConstructorFunctionBuilder';
+import { buildNamedConstructorFunction } from '../../../../../package-agnostic-utilities/constructor-function/buildNamedConstructorFunction';
 
 export type FileDependencyPathSegment = {
   tailDirectoryPath: string;
