+++ packages/voictents-and-estinants-engine/src/adapted-programs/programs/render-knowledge-graph/dependency/fileDependency.ts
@@ -1,7 +1,7 @@
 import { InMemoryOdeshin2IndexByName } from '../../../../layer-agnostic-utilities/voictent/inMemoryOdeshinVoictent2';
 import { InMemoryVoque } from '../../../../layer-agnostic-utilities/voque/inMemoryVoque';
 import { assertNotUndefined } from '../../../../package-agnostic-utilities/nil/assertNotUndefined';
-import { buildNamedConstructorFunction } from '../../../../package-agnostic-utilities/constructor-function/namedConstructorFunctionBuilder';
+import { buildNamedConstructorFunction } from '../../../../package-agnostic-utilities/constructor-function/buildNamedConstructorFunction';
 import { isNotNull } from '../../../../package-agnostic-utilities/nil/isNotNull';
 import { SimplifyN } from '../../../../package-agnostic-utilities/type/simplify';
 import { BoundedFile } from '../file/boundedFile';
