+++ packages/voictents-and-estinants-engine/src/adapted-programs/programs/render-knowledge-graph/directory/boundedDirectory.ts
@@ -1,5 +1,5 @@
 import { posix } from 'path';
-import { buildNamedConstructorFunction } from '../../../../package-agnostic-utilities/constructor-function/namedConstructorFunctionBuilder';
+import { buildNamedConstructorFunction } from '../../../../package-agnostic-utilities/constructor-function/buildNamedConstructorFunction';
 import { SimplifyN } from '../../../../package-agnostic-utilities/type/simplify';
 import { Directory } from '../../../programmable-units/file/directory';
 import { FileSystemNodeZorn } from '../../../programmable-units/file/fileSystemNode';
