+++ packages/voictents-and-estinants-engine/src/adapted-programs/programs/render-knowledge-graph/directory/directoryFact2.ts
@@ -1,6 +1,6 @@
 import { posix } from 'path';
 import { assertNotNull } from '../../../../package-agnostic-utilities/nil/assertNotNull';
-import { buildNamedConstructorFunction } from '../../../../package-agnostic-utilities/constructor-function/namedConstructorFunctionBuilder';
+import { buildNamedConstructorFunction } from '../../../../package-agnostic-utilities/constructor-function/buildNamedConstructorFunction';
 import {
   GenericZorn2Template,
   Zorn2,
