+++ packages/voictents-and-estinants-engine/src/adapted-programs/programs/render-knowledge-graph/file/fileFact2.ts
@@ -1,5 +1,5 @@
 import { InMemoryOdeshin2ListVoque } from '../../../../layer-agnostic-utilities/voictent/inMemoryOdeshinVoictent2';
-import { buildNamedConstructorFunction } from '../../../../package-agnostic-utilities/constructor-function/namedConstructorFunctionBuilder';
+import { buildNamedConstructorFunction } from '../../../../package-agnostic-utilities/constructor-function/buildNamedConstructorFunction';
 import { isNotNull } from '../../../../package-agnostic-utilities/nil/isNotNull';
 import {
   GenericZorn2Template,
