+++ packages/voictents-and-estinants-engine/src/adapted-programs/programs/render-knowledge-graph/layer/layerConfiguration.ts
@@ -1,5 +1,5 @@
 import { InMemoryOdeshin3Voque } from '../../../../layer-agnostic-utilities/voictent/inMemoryOdeshinVoictent2';
-import { buildNamedConstructorFunction } from '../../../../package-agnostic-utilities/constructor-function/namedConstructorFunctionBuilder';
+import { buildNamedConstructorFunction } from '../../../../package-agnostic-utilities/constructor-function/buildNamedConstructorFunction';
 import { SimplifyN } from '../../../../package-agnostic-utilities/type/simplify';
 import { LayerZorn } from './layerZorn';
