+++ packages/voictents-and-estinants-engine/src/adapted-programs/programs/render-knowledge-graph/partition-fact/partitionFact.ts
@@ -1,6 +1,6 @@
 import { posix } from 'path';
 import { InMemoryOdeshin3Voque } from '../../../../layer-agnostic-utilities/voictent/inMemoryOdeshinVoictent2';
-import { buildNamedConstructorFunction } from '../../../../package-agnostic-utilities/constructor-function/namedConstructorFunctionBuilder';
+import { buildNamedConstructorFunction } from '../../../../package-agnostic-utilities/constructor-function/buildNamedConstructorFunction';
 import {
   GenericZorn2Template,
   Zorn2,
